Improves boundary facet detection and cell handling
Refactors boundary facet detection for efficiency by reusing a precomputed map and using vector capacity.

Introduces `clear_neighbors` method to the `Cell` struct, allowing for the resetting of cell neighbors.
This is particularly useful in benchmarking scenarios where neighbor assignment needs to be isolated and measured.

Fixes a minor documentation issue in the README.
Adds `supercells` to the cspell dictionary to avoid false positives.
Also introduces an enum to represent the errors that can occur when finding extreme coordinates.