Improves baseline artifact lookup efficiency
Optimizes the benchmark workflow by caching baseline artifacts from recent successful workflow runs. This significantly reduces the number of API calls required to locate the correct baseline, particularly for release baselines. It also falls back
to the most recent baseline artifact if no release baseline is found.

Also, the `compare` command is executed using an `if` statement that determines whether the benchmark utils is installed as a
module or an entrypoint to avoid errors.